readd half when saving model

+++ ultralytics/engine/trainer.py
@@ -485,8 +485,8 @@ class BaseTrainer:
         ckpt = {
             "epoch": self.epoch,
             "best_fitness": self.best_fitness,
-            "model": deepcopy(de_parallel(self.model)),
-            "ema": deepcopy(self.ema.ema),
+            "model": deepcopy(de_parallel(self.model)).half(),
+            "ema": deepcopy(self.ema.ema).half(),
             "updates": self.ema.updates,
             "optimizer": self.optimizer.state_dict(),
             "train_args": vars(self.args),  # save as dict
